fix(chat): extract lead form validation logic
