+++ components/AIChat.tsx
@@ -13,7 +13,8 @@ import ReactMarkdown from 'react-markdown';
 import { getTravelAdvice } from '../services/geminiService';
 import { useLeadCapture, type SubmitLeadHookResult } from '../hooks/useLeadCapture';
 import { PassportStamp } from './ui/PassportStamp';
-import { ChatLeadForm, type LeadFinalizePayload, type LeadFinalizeResult } from './ChatLeadForm';
+import { ChatLeadForm } from './ChatLeadForm';
+import type { LeadFinalizePayload, LeadFinalizeResult } from '../lib/chat-lead-form-logic';
 import type { SubmitLeadRequest } from '../types/leadCapture';
 import { triggerHaptic } from '../utils/haptics';
 import { sanitizeAiLinkUrl } from '../lib/ai/safe-link';
